Stop re-using NthIndexCache to avoid ABA problem
I just noticed that the NthIndexCache is internally keyed on the pointer to the
selectors themselves which I think opens this up for an ABA problem, i.e. after
dropping a selector and allocating a new one at the same address but with
different content, the cache might result in incorrect matches.

Hence, I just avoided re-use completely which is certainly correct if
unnecessarily inefficient, but I fear proper re-use would need to be
user-visible.